+++ js/imageFormValidator/domElements.js
@@ -6,6 +6,7 @@ export const imageUploadForm = document.querySelector('.img-upload__form');
 export const imageUploadFormTag = document.querySelector('.text__hashtags');
 export const imageUploadFormText = document.querySelector('.text__description');
 export const imageUploadPreview = document.querySelector('.img-upload__preview img');
+export const imageUploadEffectLevel = document.querySelector('.img-upload__effect-level');
 export const imageUploadScaleControlValue = document.querySelector('.scale__control--value');
 export const imageUploadScaleControlSmaller = document.querySelector('.scale__control--smaller');
 export const imageUploadScaleControlBigger = document.querySelector('.scale__control--bigger');
