+++ js/imageFormValidator/utils.js
@@ -1,0 +1,5 @@
+export const fileReader = new FileReader();
+
+export function splitTextWithSpace(value) {
+  return value.split(/\s+/);
+}
